Add execution, orchestration, and step registry guides
Complete the Orchestration guide

Add a basic flowchart for how a scenario is structured in OpenShift CI

Change "tools" to "steps"

Typo - "a" to "an"

Type - "a" to "an"... again.

Change wording.

Complete the execution guide

Added quote blocks for phase definitions so they standout. Also modified the definitions to match the orchestration and execution guides

Adding bones of step registry guide

Complete the ref guide

Remove line

Typo

Finish up step registry guides and add them to the SUMMARY file